+++ src/Service/CodeBuild/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - AWS api-change: This release adds support for environment type WINDOWS_SERVER_2022_CONTAINER in ProjectEnvironment
 - AWS api-change: Added `us-isob-east-1` region.
+- AWS api-change: Add support for custom instance type for reserved capacity fleets
 
 ### Changed